Subtracting min(ocean_time) from ocean_time array as workaround to ParaView bug
ParaView when saving data or geometry will write a .pvd metafile with the timestep
information. For whatever reason, it will write timesteps in scientific notation
instead of as literal value. This is a problem when using modified Julian Day in
seconds or similar because there will be a loss of precision in the timestep value and
ParaView will then load multiple timesteps at once. A quick and dirty fix was to
subtract the minimum time value from the ocean_time array, so times start at 0, but
have the same dt. This won't affect ParaView's pathline generation, but is something
to be aware of.