Changed to work with new merged netcdf which is a volume instead of just a slice
Originally I had been merging the mesh netcdf with the data netcdf and only writing out the top slice with R+ncdf4. This had limitations in the ability to write out certain variables and attributes so I switched to using the NetCDF4 Python API and wrote out the entire volume. This required that the slicing now occur in the ParaView pipeline itself.